requirements.txt
Flask
openai
python-dotenv
PyPDF2
mysql-connector-python
requests
llamaapi